pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "window-art"
version = "0.1.0"
description = "A minimal Python library for live coding visual scenes using desktop windows"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Window Art Contributors" }
]
keywords = ["desktop", "windows", "live-coding", "visual", "animation", "creative-coding"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "pillow>=12.1.0",
    "pysdl2>=0.9.16",
    "pysdl2-dll>=2.28.0",
]

[project.optional-dependencies]
video = [
    "opencv-python>=4.10.0",
]
dev = [
    "pytest>=7.0",
    "mypy>=1.0",
    "opencv-python>=4.10.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.0",
    "mkdocstrings[python]>=0.24",
]

[project.urls]
Homepage = "https://github.com/willmeyers/window-art"
Repository = "https://github.com/willmeyers/window-art"

[tool.hatch.build.targets.wheel]
packages = ["window_art"]

[tool.mypy]
python_version = "3.10"
strict = true
